requirements.txt
Flask==2.3.2
PyJWT==2.7.0
python-dotenv==1.0.0
stripe==5.5.0
requests==2.31.0
gunicorn==21.2.0